pyproject.toml
[project]
name = "pixel-color"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Fabrizio", email = "fabrizio.nastri@flexup.org" }
]
dependencies = [
    "pyautogui>=0.9.54",
    "pillow>=11.0.0",
    "keyboard>=0.13.5",
    "pyperclip>=1.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pixel_color"]
